gh-439: Refactoring
